Filter default value

+++ notGimp/blurs.cpp
@@ -2,15 +2,15 @@
 
 namespace filter {
 
-	void gaussian(cv::Mat& srcfile, int8_t radius, float sigma) {
+	void gaussian(cv::Mat& srcfile, int radius, float sigma) {
 
 	}
 
-	void radial(cv::Mat& srcfile, int8_t radius, float angle) {
+	void radial(cv::Mat& srcfile, int radius, float angle) {
 
 	}
 
-	void bokeh(cv::Mat& srcfile, int8_t radius, float bladeCount) {
+	void bokeh(cv::Mat& srcfile, int radius, float bladeCount) {
 
 	}
 }
